PMM-7492 show/hide fields based on auth type
