+++ public/app/percona/settings/components/Communication/Email/Email.types.ts
@@ -1,0 +1,7 @@
+import { EmailSettings } from '../../../Settings.types';
+import { LoadingCallback } from '../../../Settings.service';
+
+export interface EmailProps {
+  settings: EmailSettings;
+  updateSettings: (body: any, callback: LoadingCallback) => void;
+}
